requirements.txt
# Core dependencies for obs-tools
# Minimal by design - most functionality works with stdlib only

# macOS-specific dependencies (optional, installed automatically on macOS)
pyobjc>=8.0; sys_platform == "darwin"
pyobjc-framework-EventKit>=8.0; sys_platform == "darwin"

# Optional optimization dependencies
scipy>=1.5.0  # For Hungarian algorithm in bipartite matching (falls back to pure Python if unavailable)

# Optional validation dependencies  
jsonschema>=3.0.0  # For schema validation (falls back to basic validation if unavailable)

# Development dependencies (install with pip install -e .[dev])
# pytest>=6.0
# pytest-cov>=2.10
# black>=21.0
# mypy>=0.900